+++ packages/sidesheet/src/lib/types/types.ts
@@ -1,0 +1,8 @@
+export type ReplaceFunction = (
+	newComp: (el: HTMLDivElement, replace: Frame) => Promise<VoidFunction> | VoidFunction
+) => Promise<VoidFunction>;
+
+export type Frame = {
+	replace: ReplaceFunction;
+	unmount: () => void;
+};