+++ packages/workspace-fusion-modules/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion-modules@0.1.3...@equinor/workspace-fusion-modules@0.1.4) (2022-11-15)
+
+**Note:** Version bump only for package @equinor/workspace-fusion-modules
+
 ## [0.1.3](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion-modules@0.1.2...@equinor/workspace-fusion-modules@0.1.3) (2022-11-14)
 
 **Note:** Version bump only for package @equinor/workspace-fusion-modules